pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "job-detect"
version = "0.1.0"
description = "求职决策 CLI — 不是搜索工具，是决策工具"
requires-python = ">=3.10"
dependencies = [
    "typer>=0.12",
    "rich>=13",
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "zhipuai>=2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "ruff>=0.5",
]

[project.scripts]
job-detect = "src.cli.main:app"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]
